Level Design (Part 1)  | |  Week 25: Day 6 (2/26) --- Post 2
I gave each of the terrain assets in the 21 levels their own Data so that one can be altered while the other remain the same, this was a weird issue but one that got solved pretty quickly. I stored the data, along with the map brushes, in their respected folders under the Terrain section. Time to start actually applying the maps into the terrain. All I really need to do is print the maps in at Brush Strength 0.25, Brush Size 350, then lower the terrain by -5 units. After doing all of this, I can smooth out the edges of the terrain and get a prettier result. The exact values may change but the core principles are the same.