Level Design (Part 2)  | |  Week 29: Day 3 (3/23)  --- Post 1
Welp, its been nearly a month since I worked on this project. I decided to temporarily put a halt to Trench Fever to work on a project to help out the people in Ukraine, this ultimately culminated into the release of "Chess for Ukraine", a chess game released on Itch.io that serves as a fundraiser. All proceeds will be donated to the GlobalGiving "Ukraine Crisis Relief Fund". I felt I needed to do something, and I'm proud to have my official "FIRST PUBLISHED GAME" be a fundraiser to help innocent civilians during time of war.

Now that the game is already up on Itch.io and doing pretty good for itself, I decide it's about time to return to my Original First Game. It's been a while and now that I only have three days a week to work on the project, I need to really get an idea of what I have left to do an get it done. I don't want this project to be sitting around and never getting completed, but I also don't want to spend a great deal of time getting everything working. I still want this game to be out by my 20th birthday, if that is even possible. I've made compromises before to ensure this project will get out on time, and I will do so again to ensure everything gets released in a timely fashion. I want this to be a good game, as good as I can make it given the time, but I've also grown rather distant with the project even before the Ukraine crisis. I just want this project finished and published, make sure everything is in working order before release and move on to the next great idea.

I really don't have much left to do with this game, I will list below the few remaining features I need to complete before I can get this ship sailing:
     1. Main Menu
     2. Pause Menu
     3. Complete all Levels
     4. Get all Exposition Scenes
     5. Finish the Story
     6. Update Graphics if Needed
     7. Update Audio if Needed

Alright, now lets get to it!