Level Design (Part 2) | | Week 29: Day 4 (3/24) --- Post 1
I went ahead with my plans to finish the level design by painting the corresponding maze patterns into the Terrain, slightly lowing the terrain, and creating a plane to act as the ground. Everything is looking pretty neat thus far. I just need to work on placing all of the props, items, enemy spawners, etc. I'll need to get done with this pretty quickly if I want to get the game finished on time!