Level Design (Part 3) | | Week 30: Day 2 (3/29) --- Post 1
I improved the spawning system so that enemies will only spawn when their spawner gameobjects are not currently being looked at by the player's camera. Everything is working pretty well so far and I'm proud to say my game is starting to come together. I might mess with some audio today to really try and bring my game to life. I want there to be an initial sound effect when the enemy spawns, something sounding like a man jumping from a high place, to reflect a soldier jumping into the trenches. I also want to create a droplet sound effect when the blood droplets hit the ground to spawn the splash effect. And finally, I NEED an effect for when the enemy projectiles collide with a object. At the current moment, they do nothing. I want a little effect and a sound effect to really give impact to it. I'll continue working to try and get the game working as smoothly as possible. Hopefully, I can get this ship sailing soon!!!