Level Design (Part 4) | | Week 35: Day 3 (5/4) --- Post 2
I made a good bit of changes to the game to help improve the overall product. I finally decided to create the Main Menu with all of the basic buttons preforming just nicely. I need to get the subMenus for the Levels, Encyclopedia and Credits taken care of sometime. I created a basic fade in and out text system for the Exposition scenes, which will add a nice bit of flavor to the game and allow me to tell a story, albeit a basic one. I also went into the player prefab and made it so that each weapon does not automatically reload when depleted, this was only done for testing purposes. I need to make it known that the player is only meant to have limited ammo during their run per level. I need to sit down and think about how I want to carry out the Encyclopedia, Level Selection and Credits. Do I want them to be subMenus of the MainMenu scene or should they be separate scenes to make things more organized? Anyways, good job!!!