Level Design (Part 4) | | Week 35: Day 5 (5/6) --- Post 1
I went through all of the levels to make sure everything is in working order. I was able to find all of the exit doors within 1 minute or less, although I'll need to see with playtesters if the maze is too difficult. I modified the DeathRestart script so that if the player dies in the Finale, they will be sent to the C5 Ending instead of restarting the levels.