Level Design (Part 5) | | Week 36: Day 1 (5/9) --- Post 1
I was pretty busy today with work and visiting Grandpa in the hospital, so I didn't have much time to work on this video game. I went ahead and created a title for my Encyclopedia scene simply because I needed one, and that was it. I'll have much more planned for tomorrow when I start expanding my Encyclopedia scene and fleshing out the levels.