Level Design (Part 5) | | Week 36: Day 3 (5/11) --- Post 1
Looks like I forgot to make a post yesterday even though I got a good bit done. Yesterday, I created the foundation for the information pages for the History and Weapons category in the Encyclopedia, this includes a title, description and corresponding images. Today, I made these same changes to the Enemies category.