Level Design (Part 5) | | Week 36: Day 4 (5/12) --- Post 1
Try and get my Enemies to spawn properly on all of my other levels. I introduced the new spawnPoints to level 2 and placed them throughout the level. But for some odd reason, the enemies are stuck in one spot unable to move. I baked the NavMesh but apparently that isn't working. I know I've faced this problem before, I just don't know how to solve it. I'm going to be doing some research on how to get this solved.