Level Design (Part 5) | | Week 36: Day 5 (5/13) --- Post 1
Finally figured out what to do! After I import the new EnemySpawnPoints, I need to select the Ground plane and bake a NavMesh onto it. The main reason the enemies hadn't been moving properly is because I've been selecting the Terrain or something else entirely when building the NavMesh, resulting in the enemies being spawned on a surface they can't walk on. I might actually go ahead and do this for every level right quick before I add in the new spawners since it's much easier to accomplish.