Level Design (Part 5) | | Week 36: Day 5 (5/13) --- Post 2
I went into all the other levels and generated NavMeshes for the ground planes. I still need to import the new and improved EnemySpawnpoints so that the enemies can actually spawn into the levels. I created a new folder under each chapter that holds all of the NavMesh information for the levels. Everything looks like it's finally starting to come together!