Level Design (Part 5) | | Week 36: Day 5 (5/13) --- Post 3
I deleted the original EnemySpawnpoints in the respected levels and replaced them with a copy & paste version of the ones from the original level. I still need to put the respected spawnPoints in the EnemySpawner slots and place them in the level where it would make sense. This might take a hot minute!