Level Design (Part 5) | | Week 36: Day 5 (5/13) --- Post 5
I finally went through all 21 of my levels and got the EnemySpawnpoints positioned in the right spots, the only other things I really need to be concerned with in my level design going forward is to place all of my obstacles and HealthKits. After that, it really is a matter of taking what's already created and improving upon it! I can't believe I'm almost there!!!!!