Level Design (Part 6) | | Week 37: Day 4 (5/19) --- Post 4
I made a few minor alterations to 2 levels. I playtested each of them to make sure everything was working as intended. The game honestly plays great although there is definitely a big that needs to be improved. I found that the final level (C4L5) is pretty difficult, might be a well deserved victory when I finish it. It's a nice challenge. I definitely need to make several changes, including fixing the sound issue when completing the level and solving the GasEffect bug.

I also discovered a very strange bug on C2L2 I believe. I was sprinting backwards towards the exit door while firing my Lewis Gun. Evidently I missed the exit door and moved into the wall next to it, going through it without collision and exited the playable area. I have not been able to replicate this glitch yet but I might just keep it in, it would be an interesting speedrunning tactic.