Level Design (Part 7) | | Week 38: Day 2 (5/24) --- Post 1
Didn't make a post last weekend because I was tired from working, and didn't yesterday because I was watching Prehistoric Planet for the first time (le Epic!!!). I'm on a mission right now to try and make each level unique in their own way. I just made a Chapter one level, nicknamed "Horde", a solely rat based map. I also went into all of the levels and changed the Enemy capacity, so that there will be a higher maximum of enemies in the later chapters.