Level Design (Part 7) | | Week 38: Day 2 (5/24) --- Post 2
I added in a naming system to each of the levels so that it will shortly display the name and chapter of the map currently being played. I want to play around with more zanny map ideas such as having levels that focus on a particular threat, maybe that could be the way of introducing new elements. I'll have to sit back and ponder about them.