Level Design (Part 7) | | Week 38: Day 3 (5/25) --- Post 2
Well, I did something that is known in the industry as an "opsie". I meddled with the terrain in one of my test levels to figure out if the gas cloud is working fine there. What I didn't realize is that that level's terrain data is connected to the terrain data of the first level, so I accidently erased my first level's terrain. I tried undoing to get it back but it was already too late, it was gone. I've successfully replaced the trench map into the terrain and got everything lined back up, including re-baking the navigation data for the enemies so they don't just collide with walls. No I'm doing to try and figure out what is going wrong with my Gas clouds once and for all!