Level Design (Part 7) | | Week 38: Day 4 (5/26) --- Post 1
I think I found the problem. The actual gas effect in the Player HUD prefab is missing the material shader that's present on the gameobject in the testing levels. The trick is, I have no clue how to get that to appear in the standard levels. Eventually I'll get this figured out, I'm just glad I finally diagnosed the problem!