Level Design (Part 8) | | Week 39: Day 2 (5/31) --- Post 1
I copy and pasted an effects object into the 21 levels, and replaced the gas_damage effect on the player in the first level with the new pasted one, and they are working fine. However, I've encountered trouble with the gassed_effect. I can't find anywhere where this effect would be applied. It is the effect that appears when the player is wearing the gasMask and has walked into the gasCloud. I can't for the life of me try and figure out where it is. It might actually be on the gasClouds themselves, give me a second...