Finishing Touches (Part 1) | | Week 40: Day 2 (6/7) --- Post 1
I'm entering a new and final stage in my game development process I've dubbed the "Finishing Touches" phase. The remainder of my tasks are really just polishing work to ensure maximum enjoyment in player experience, nothing really major or "game-breaking" (har-har). Today was the first time I've ever built out my project as an actual game, and I was really impressed with the results, it definitely helped put everything into perspective and allow myself to realize how close I am to actually finishing my project. This was supposed to be my final week working on this game, and it is unfortunate I got this far behind even if my sights were for doing the greater good.

Anyways, here is an impressed list of things I need to accomplish during my new polishing phase as to actually get this game up and running:
   1. Fix GasEffect Issue in C3L5
   2. Move Door in C3L5
   3. Place in Background Objects in Levels
   4. Polish Menu UI
   5. Fix Next Button Text Issue
   6. Fill out information in Encyclopedia
   7. Fill out Credits
   8. Fill out Exposition
   9. Soften Terrain edges in Levels
   10. Add in Music in each level
   11. Fix the EndDoor Audio Issue
   12. Straighten all of the EndDoors