Finishing Touches (Part 1) | | Week 40: Day 5 (6/10) --- Post 1
We saw Jurassic World: Dominion yesterday in theaters. I'm currently mixed at the moment, it just felt like it had weird pacing and a whole bunch of stories that lead nowhere. I've never felt this way about a Jurassic Park film, and it makes me a little disappointed. We'll see if the movie grows on me...

I fleshed out the main menu of the game, I'm going to continue working on the other menus to get everything looking right.