Finishing Touches (Part 3) | | Week 42: Day 2 (6/21) --- Post 1
I made the tough decision to not include the Encyclopedia section in the game, it would be too much work and get finished and given I don't have a lot of time left to get this game up and running, it just doesn't seem like that much of a priority. What I need to do is work out all of the little kinks in the game and get everything looking nice and beautiful (or as pretty as I can manage), adding an Encyclopedia section which would only be used by a tiny minority of players doesn't feel like a very pressing matter. Perhaps I could add it back in sometime in the future, but it's not an issue at the moment.