Finishing Touches (Part 3) | | Week 42: Day 4 (6/23) --- Post 1
I couldn't do much today since it was spent watching a movie and traveling with family. I can't believe I don't have much left to complete before this game can be considered done in my opinion. I really don't have much left that needs to be done. I need to focus on the music at the moment and worry about the polishing up later, music in the levels is a priority at the moment!