Finishing Touches (Part 4) | | Week 43: Day 2 (6/28) --- Post 2
I added all of the appropriate songs to the levels and exposition scenes, I also double checked each of them to make sure I didn't double placed the same music. I placed songs depending on the context of the story at the moment, each choice was made to enhance the narrative (what little their was). Now I need to get the music to accomplish two things in the standard levels.

1. To disappear when the player is dead
2. To disappear when the level is completed.

This should take long at all.