Finishing Touches (Part 4) | | Week 43: Day 2 (6/28) --- Post 3
I modified the scripts so that the music will disappear if the player either dies or wins the level. This is a really neat system that I think will improve the experience. Now I want to focus on getting the music in the exposition scenes working correctly, where they will gradually low in volume before transitioning to the next level.