Finishing Touches (Part 4) | | Week 43: Day 3 (6/29) --- Post 4
I added special tags for the BombManager and ArtilleryManager gameObjects and applied them to each of the instances in the levels. I also modified the WinCondition script so that these objects will be found and deleted as needed. I went around through the different chapters and made alterations to the lighting and fog so that each chapter has a unique feel to it.