Finishing Touches (Part 5) | | Week 44: Day 2 (7/5) --- Post 1
I created 4 original images for the Win Screen of the 20 levels that possess such. They've already been added, as well as the titles changed to reflect the nature of the chapter. I slightly modified the Time text and added a PAR time, which would let players know what time they should try and beat the level at.