Finishing Touches (Part 5) | | Week 44: Day 4 (7/7) --- Post 1
I went through each of the levels and updated the PAR time to match that of the respected level. I also discovered that the second paragraph animation and music cut off was not right in the Exposition. I went ahead and fixed those up. I really don't have much left to complete with this game but now I'm thinking about if I should charge anyone for playing, because I really don't see myself paying money to play this game.