Finishing Touches (Part 5) | | Week 44: Day 4 (7/7) --- Post 2
I changed a mismatched issue with the PAR time and the recorded time, as well as change the Next button a little bit. My theory as to why the Next text disappears when the game is built is because there is something lost in translation and the text becomes too big for the button, so I made the text smaller. Hopefully this will work.