Finishing Touches (Part 5) | | Week 44: Day 4 (7/7) --- Post 4
Added a dead Frenchmen to C1L5 to add more detail to the level, I also littered the C3 and C4 levels with dead British soldiers to really show that this is a British battalion. There are some dead Brits in C2L5 as well to show that you are at their telegram communication station. I completely littered C3L5 with over seventy British bodies and several Stormtrooper bodies to show that this was a massacre.