Finishing Touches (Part 6) | | Week 45: Day 2 (7/12) --- Post 6
Got the terrain pixelated as well as the gun shells, I then made alterations to the gas mask and weapons to make them more pixelated. Now I just need to look back through the level's names and make any necessary changes and figure out what is going on with the endDoors not detecting collision properly.