Finishing Touches (Part 6) | | Week 45: Day 4 (7/14) --- Post 1
I made some modifications to the Credits, Make the level names smaller in the Load Level scene so that they don't cut off, fixed spelling mistakes in the Exposition, lengthened Exposition music and got the first level of every chapter to begin with one of the newly introduced weapons. I really can't think of anything else to add to this game, I might be finished. I'll build it out and play around with it to see if everything is playing properly then go from there.